frontend: Extract ThemeChangeButton from App
